Add CSV parser
An efficient workflow for entering results is to type the results into
a spreadsheet (taking advantage of autocomplete for names) and then
manipulate this into the expected format for display and publication.

This adds a limited but functional parser of these data-entry optimised
CSV files so the rest of the functionality of the "BAR results parser"
can be used for generating publishable and validated results.