2012-Spring: Add the results of the Apple Prize
I am trying something a little different to what has appeared in the
results in previous years.  I am listing all the people who shot for the
apple and just marking the winner with a score of 1 rather than 0.  This
will not yet appear in the results until I add support for competitions
of type "Apple" to the parser but I think it should make the results a
bit more interesting once this has been added.