2012-Spring: Remove duplicate score for Neil Jones in the Eröff
There are two scores in the scorebook and two cards entered.  Neither
card has the re-entry box ticked.  Have removed the second of these
assuming that he really did re-shoot it having forgotten he had shot it
before.

Either score 97 or 119 is a poor score for Neil so this is slightly
suspicious but there's little else that can be done at this point.